backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pandas==2.2.3
python-multipart==0.0.20
openpyxl==3.1.5
httpx==0.28.1
openai==1.70.0
python-docx==1.1.2
matplotlib==3.9.4
